int instead of size_t

+++ src/utilization/gpu/GPUUtilizationWorker.cpp
@@ -5,6 +5,6 @@
 void GPUUtilizationWorker::receiveData() {
     auto gpuUtilization = NVSMIParser::getGPUUtilization();
 
-    for (size_t i = 0; i < gpuUtilization.size(); i++)
+    for (int i = 0; i < gpuUtilization.size(); i++)
         udata[i].level = gpuUtilization[i];
 }
